front: move NGE timetable loading to osrdToNge.ts
Moves macro editor details out of ScenarioContent.

This function will grow soon.
